% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcuityView_function.R
\name{fft_matrix_shift}
\alias{fft_matrix_shift}
\title{FFTMatrixShift}
\usage{
fft_matrix_shift(input_matrix, dim = -1)
}
\description{
This function rearranges the output of the FFT by moving the zero frequency component to the center
fft_matrix_shift()
}
