% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcuityView.R
\name{fft_matrix_shift}
\alias{fft_matrix_shift}
\title{FFTMatrixShift}
\usage{
fft_matrix_shift(input_matrix, dim = -1)
}
\arguments{
\item{input_matrix}{the output of an FFT}

\item{dim}{-1 gives the correct matrix shift for the AcuityView function}
}
\description{
This function rearranges the output of the FFT by moving the zero frequency component to the center
}
